Somme changements, renamed testIfMinimal, and code sum all power
I corrected a logic error code
The inside loop,
indexListMinimal was used incorrectly.
the goal was to use indexListMinimal[0]
for index listMinimal,
because indexListMinimal is a table and not a number.

renamed testIfMinimal() to power()

now power calculate
the multiplication of the minimum

for each line with
rl.on('line', (line) => {
now the code sum all the power.